✨ [Feat] Model 타입에서 날짜 포맷해주는 프로퍼티 추가
- 구조체명 NoteItem -> NoteData로 수정
