--- CloudNotes/CloudNotes/NoteItem.swift
@@ -1,20 +1,0 @@
-//
-//  NoteItem.swift
-//  CloudNotes
-//
-//  Created by 배은서 on 2021/06/01.
-//
-
-import Foundation
-
-struct NoteItem: Decodable {
-    let title: String
-    let body: String
-    let lastModified: Int
-    
-    private enum CodingKeys: String, CodingKey {
-        case title, body
-        case lastModified = "last_modified"
-    }
-    
-}
